+++ trade-engine/src/test/java/com/icoin/trading/tradeengine/saga/PortfolioCommandHandlerTest.java
@@ -50,6 +50,8 @@ import org.axonframework.test.Fixtures;
 import org.junit.Before;
 import org.junit.Test;
 
+import java.math.BigDecimal;
+
 /**
  * @author Jettro Coenradie
  */
@@ -87,10 +89,14 @@ public class PortfolioCommandHandlerTest {
     public void testAddItemsToPortfolio() {
         AddItemsToPortfolioCommand command = new AddItemsToPortfolioCommand(portfolioIdentifier,
                 orderBookIdentifier,
-                100);
+                BigDecimal.valueOf(100));
         fixture.given(new PortfolioCreatedEvent(portfolioIdentifier, userIdentifier))
                 .when(command)
-                .expectEvents(new ItemsAddedToPortfolioEvent(portfolioIdentifier, orderBookIdentifier, 100));
+                .expectEvents(
+                        new ItemsAddedToPortfolioEvent(
+                                portfolioIdentifier,
+                                orderBookIdentifier,
+                                BigDecimal.valueOf(100)));
     }
 
     @Test
@@ -98,7 +104,7 @@ public class PortfolioCommandHandlerTest {
         ReserveItemsCommand command = new ReserveItemsCommand(portfolioIdentifier,
                 orderBookIdentifier,
                 transactionIdentifier,
-                200);
+                BigDecimal.valueOf(200));
         fixture.given(new PortfolioCreatedEvent(portfolioIdentifier, userIdentifier))
                 .when(command)
                 .expectEvents(new ItemToReserveNotAvailableInPortfolioEvent(portfolioIdentifier, orderBookIdentifier, transactionIdentifier));
@@ -109,15 +115,15 @@ public class PortfolioCommandHandlerTest {
         ReserveItemsCommand command = new ReserveItemsCommand(portfolioIdentifier,
                 orderBookIdentifier,
                 transactionIdentifier,
-                200);
+                BigDecimal.valueOf(200));
         fixture.given(new PortfolioCreatedEvent(portfolioIdentifier, userIdentifier),
-                new ItemsAddedToPortfolioEvent(portfolioIdentifier, orderBookIdentifier, 100))
+                new ItemsAddedToPortfolioEvent(portfolioIdentifier, orderBookIdentifier, BigDecimal.valueOf(100)))
                 .when(command)
                 .expectEvents(new NotEnoughItemsAvailableToReserveInPortfolio(portfolioIdentifier,
                         orderBookIdentifier,
                         transactionIdentifier,
-                        100,
-                        200));
+                        BigDecimal.valueOf(100),
+                        BigDecimal.valueOf(200)));
     }
 
     @Test
@@ -125,11 +131,15 @@ public class PortfolioCommandHandlerTest {
         ReserveItemsCommand command = new ReserveItemsCommand(portfolioIdentifier,
                 orderBookIdentifier,
                 transactionIdentifier,
-                200);
+                BigDecimal.valueOf(200));
         fixture.given(new PortfolioCreatedEvent(portfolioIdentifier, userIdentifier),
-                new ItemsAddedToPortfolioEvent(portfolioIdentifier, orderBookIdentifier, 400))
+                new ItemsAddedToPortfolioEvent(portfolioIdentifier,
+                        orderBookIdentifier,
+                        BigDecimal.valueOf(400)))
                 .when(command)
-                .expectEvents(new ItemsReservedEvent(portfolioIdentifier, orderBookIdentifier, transactionIdentifier, 200));
+                .expectEvents(new ItemsReservedEvent(portfolioIdentifier,
+                        orderBookIdentifier, transactionIdentifier,
+                        BigDecimal.valueOf(200)));
     }
 
     @Test
@@ -138,15 +148,19 @@ public class PortfolioCommandHandlerTest {
                 new ConfirmItemReservationForPortfolioCommand(portfolioIdentifier,
                         orderBookIdentifier,
                         transactionIdentifier,
-                        100);
+                        BigDecimal.valueOf(100));
         fixture.given(new PortfolioCreatedEvent(portfolioIdentifier, userIdentifier),
-                new ItemsAddedToPortfolioEvent(portfolioIdentifier, orderBookIdentifier, 400),
-                new ItemsReservedEvent(portfolioIdentifier, orderBookIdentifier, transactionIdentifier, 100))
+                new ItemsAddedToPortfolioEvent(portfolioIdentifier,
+                        orderBookIdentifier,
+                        BigDecimal.valueOf(400)),
+                new ItemsReservedEvent(portfolioIdentifier,
+                        orderBookIdentifier, transactionIdentifier,
+                        BigDecimal.valueOf(100)))
                 .when(command)
                 .expectEvents(new ItemReservationConfirmedForPortfolioEvent(portfolioIdentifier,
                         orderBookIdentifier,
                         transactionIdentifier,
-                        100));
+                        BigDecimal.valueOf(100)));
     }
 
     @Test
@@ -155,60 +169,69 @@ public class PortfolioCommandHandlerTest {
                 new CancelItemReservationForPortfolioCommand(portfolioIdentifier,
                         orderBookIdentifier,
                         transactionIdentifier,
-                        100);
+                        BigDecimal.valueOf(100));
         fixture.given(new PortfolioCreatedEvent(portfolioIdentifier, userIdentifier),
-                new ItemsAddedToPortfolioEvent(portfolioIdentifier, orderBookIdentifier, 400),
-                new ItemsReservedEvent(portfolioIdentifier, orderBookIdentifier, transactionIdentifier, 100))
+                new ItemsAddedToPortfolioEvent(
+                        portfolioIdentifier,
+                        orderBookIdentifier,
+                        BigDecimal.valueOf(400)),
+                new ItemsReservedEvent(portfolioIdentifier, orderBookIdentifier, transactionIdentifier,
+                        BigDecimal.valueOf(100)))
                 .when(command)
                 .expectEvents(new ItemReservationCancelledForPortfolioEvent(portfolioIdentifier,
                         orderBookIdentifier,
                         transactionIdentifier,
-                        100));
+                        BigDecimal.valueOf(100)));
     }
 
     /* Money related test methods */
     @Test
     public void testDepositingMoneyToThePortfolio() {
-        DepositCashCommand command = new DepositCashCommand(portfolioIdentifier, 2000l);
+        DepositCashCommand command = new DepositCashCommand(portfolioIdentifier, BigDecimal.valueOf(2000l));
         fixture.given(new PortfolioCreatedEvent(portfolioIdentifier, userIdentifier))
                 .when(command)
-                .expectEvents(new CashDepositedEvent(portfolioIdentifier, 2000l));
+                .expectEvents(new CashDepositedEvent(portfolioIdentifier, BigDecimal.valueOf(2000l)));
     }
 
     @Test
     public void testWithdrawingMoneyFromPortfolio() {
-        WithdrawCashCommand command = new WithdrawCashCommand(portfolioIdentifier, 300l);
-        fixture.given(new PortfolioCreatedEvent(portfolioIdentifier, userIdentifier), new CashDepositedEvent(portfolioIdentifier, 400))
+        WithdrawCashCommand command = new WithdrawCashCommand(portfolioIdentifier,
+                BigDecimal.valueOf(300l));
+        fixture.given(new PortfolioCreatedEvent(portfolioIdentifier, userIdentifier),
+                new CashDepositedEvent(portfolioIdentifier, BigDecimal.valueOf(400)))
                 .when(command)
-                .expectEvents(new CashWithdrawnEvent(portfolioIdentifier, 300l));
+                .expectEvents(new CashWithdrawnEvent(portfolioIdentifier, BigDecimal.valueOf(300l)));
     }
 
     @Test
     public void testWithdrawingMoneyFromPortfolio_withoutEnoughMoney() {
-        WithdrawCashCommand command = new WithdrawCashCommand(portfolioIdentifier, 300l);
-        fixture.given(new PortfolioCreatedEvent(portfolioIdentifier, userIdentifier), new CashDepositedEvent(portfolioIdentifier, 200))
+        WithdrawCashCommand command = new WithdrawCashCommand(portfolioIdentifier, BigDecimal.valueOf(300l));
+        fixture.given(new PortfolioCreatedEvent(portfolioIdentifier, userIdentifier),
+                new CashDepositedEvent(portfolioIdentifier, BigDecimal.valueOf(200)))
                 .when(command)
-                .expectEvents(new CashWithdrawnEvent(portfolioIdentifier, 300l));
+                .expectEvents(new CashWithdrawnEvent(portfolioIdentifier, BigDecimal.valueOf(300l)));
     }
 
     @Test
     public void testMakingMoneyReservation() {
         ReserveCashCommand command = new ReserveCashCommand(portfolioIdentifier,
                 transactionIdentifier,
-                300l);
-        fixture.given(new PortfolioCreatedEvent(portfolioIdentifier, userIdentifier), new CashDepositedEvent(portfolioIdentifier, 400))
+                BigDecimal.valueOf(300l));
+        fixture.given(new PortfolioCreatedEvent(portfolioIdentifier, userIdentifier),
+                new CashDepositedEvent(portfolioIdentifier, BigDecimal.valueOf(400)))
                 .when(command)
-                .expectEvents(new CashReservedEvent(portfolioIdentifier, transactionIdentifier, 300l));
+                .expectEvents(new CashReservedEvent(portfolioIdentifier, transactionIdentifier, BigDecimal.valueOf(300l)));
     }
 
     @Test
     public void testMakingMoneyReservation_withoutEnoughMoney() {
         ReserveCashCommand command = new ReserveCashCommand(portfolioIdentifier,
                 transactionIdentifier,
-                600l);
-        fixture.given(new PortfolioCreatedEvent(portfolioIdentifier, userIdentifier), new CashDepositedEvent(portfolioIdentifier, 400))
+                BigDecimal.valueOf(600l));
+        fixture.given(new PortfolioCreatedEvent(portfolioIdentifier, userIdentifier),
+                new CashDepositedEvent(portfolioIdentifier, BigDecimal.valueOf(400)))
                 .when(command)
-                .expectEvents(new CashReservationRejectedEvent(portfolioIdentifier, transactionIdentifier, 600));
+                .expectEvents(new CashReservationRejectedEvent(portfolioIdentifier, transactionIdentifier, BigDecimal.valueOf(600)));
     }
 
     @Test
@@ -216,10 +239,12 @@ public class PortfolioCommandHandlerTest {
         CancelCashReservationCommand command = new CancelCashReservationCommand(
                 portfolioIdentifier,
                 transactionIdentifier,
-                200l);
-        fixture.given(new PortfolioCreatedEvent(portfolioIdentifier, userIdentifier), new CashDepositedEvent(portfolioIdentifier, 400))
+                BigDecimal.valueOf( 200l));
+        fixture.given(new PortfolioCreatedEvent(portfolioIdentifier, userIdentifier),
+                new CashDepositedEvent(portfolioIdentifier, BigDecimal.valueOf(400)))
                 .when(command)
-                .expectEvents(new CashReservationCancelledEvent(portfolioIdentifier, transactionIdentifier, 200l));
+                .expectEvents(new CashReservationCancelledEvent(portfolioIdentifier, transactionIdentifier,
+                        BigDecimal.valueOf(200l)));
     }
 
     @Test
@@ -227,9 +252,11 @@ public class PortfolioCommandHandlerTest {
         ConfirmCashReservationCommand command = new ConfirmCashReservationCommand(
                 portfolioIdentifier,
                 transactionIdentifier,
-                200l);
-        fixture.given(new PortfolioCreatedEvent(portfolioIdentifier, userIdentifier), new CashDepositedEvent(portfolioIdentifier, 400))
+                BigDecimal.valueOf(200l));
+        fixture.given(new PortfolioCreatedEvent(portfolioIdentifier, userIdentifier),
+                new CashDepositedEvent(portfolioIdentifier, BigDecimal.valueOf(400)))
                 .when(command)
-                .expectEvents(new CashReservationConfirmedEvent(portfolioIdentifier, transactionIdentifier, 200l));
+                .expectEvents(new CashReservationConfirmedEvent(portfolioIdentifier, transactionIdentifier,
+                        BigDecimal.valueOf(200l)));
     }
 }
